pyproject.toml
[project]
name = "discord-automations-bots"
version = "0.1.0"
description = "Discord bots for automation private tasks"
authors = [
    {name = "MarcinParda",email = "marcin98parda@gmail.com"}
]
readme = "README.md"
requires-python = "<4,>=3.10"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "notion-client (>=2.3.0,<3.0.0)",
    "groq (>=0.18.0,<0.19.0)",
    "discord-py (>=2.4.0,<3.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
